More work on NAOMI support.
This commit adds a NAOMI-specific linker script. I'm actually not positive this is needed, but I couldn't find any NAOMI binaries that were linked to an earlier start address than 8c020000, so I'm just going with that for now to be safe until I can be sure it is or is not a problem.
Also, this commit moves the priority of dbgio_null above dbgio_fb so that we never start writing to the framebuffer early on if any of the other dbgio setups cannot be used.