genexports: Avoid redefining functions as different symbols
The export.txt files now support a way to specify that a particular
header should be included. For instance, to include kos.h:
include kos.h

This means that the functions that need to be exported can be fetched
from their respective headers, and they do not have to be referenced as
"extern unsigned int", which was making KallistiOS impossible to compile
with LTO.

The current export.txt files were updated with the right include files,
and the functionality in the genexports.sh script to add a
"extern unsigned int" for each symbol was removed.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>