init: Add optional callback set up with KOS_INIT_EARLY()
If supplied to the KOS_INIT_EARLY() macro, this function will be called
very early in the boot process.

This is used in my case to copy data appended to the raw binary to its
final location in RAM; the advantage vs. romdisks, is that the data
appended to the raw binary will eventually be reused by the .bss section
and the heap.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

---
v2: Provide and document KOS_INIT_EARLY() macro