dc: init: Move __kos_init_early_fn to .data section
The __kos_init_early_fn pointer is initialized to NULL, which means that
unless specified otherwise, it ends up in the .bss section.

However, its value is checked *before* the .bss section is cleared;
which means that it might contain a garbage value, and the init code
would then branch to a random address in memory.

Address this issue by forcing the value to be in the .data section.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>