Modified thd_each() to support early exit + retval
It is often the case that iteration over a thread is doing a search for
a particular thread or group of threads, allowing iteration to halt
early. It is also the case that frequently the user may wish to return
different values from their callback depending on the results.

Presently we take a user callback whose return value is simply ignored
and we return a "result code" that is hardcoded to always be zero.

- Modified thd_each() to cease iteration early upon encountering a
  non-zero return value from the user's callback
- Modified thd_each() to return the callback's return value