Garbage-collect network stack (#329)
* Disable exports by default

Add INIT_EXPORT init flag. When set (with the KOS_INIT_FLAGS macro), the
KOS symbols will be exported. The default is to not export them.

This helps reduce the size of the executables, as the vast majority of
them don't need the symbol export feature, by telling the compiler that
it's okay to garbage-collect them.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* Garbage-collect the network stack when unused

Make it so that the network stack (and BBA and LAN adapter code)
is garbage-collected when a KOS program has been compiled without the
INIT_NET flag (and doesn't use any of the symbols).

This really drops the size of the compiled binaries. For instance, the
dreamcast "hello" example dropped from 284 KiB to 156 KiB (both
stripped).

Signed-off-by: Paul Cercueil <paul@crapouillou.net>