Modularizing KOS, my take (#360)
* cdefs.h: Add and use __weak, __used, likely() and unlikely()

Add __weak and __used tokens, and use them in the places were the
corresponding GCC attributes were used.

Add likely() and unlikely() macros, which are yet to be used anywhere.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* fs_romdisk: Make init/shutdown functions void-typed

They would always return 0, and the calling code would never check the
error code anyway. So they might just as well not return any value.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* exports: Make export_init() and nmmgr_init() void-typed

Their potential errors (which in practice will never happen) were
ignored by the init code. Therefore it's just simpler to make these
functions void-typed.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* maple: Make init functions void-typed

The calling code would never check the error code, so they might just as
well not return any value.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* fs_romdisk: Make builtin romdisk mount function void-typed

The potential error codes is ignored by the init code anyway, so it can
be converted to void-typed.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* Make '__kos_init_flags' const

It does not need to be mutable and the user shouldn't update it at
runtime. Use this occasion to also update from using the deprecated
uint32 type to uint32_t.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* Add and use new macros for init/shutdown functions

The KOS_INIT_FLAG_EXPORT() is supposed to be used within
KOS_INIT_FLAGS(). It will override the weak symbol associated with the
function with a strong one, that points to this function if the flag is
set, or is NULL if the flag is cleared.

The KOS_INIT_FLAG_WEAK() should be used to declare the weak symbol for
the init/shutdown function.

Finally, the init/shutdown function can be called with
KOS_INIT_FLAG_CALL(). If the flag has not been set, the function will
not be called, and will be automatically garbage-collected by the
compiler.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* Add support for arch-specific init flags

This will later allow programs to pick which parts of the hardware
should be initialized by KallistiOS.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* Add init flags for all Maple drivers

Add init flags for all Maple drivers, so that a program can configure
which peripherals should be supported by KallistiOS (the default being
all of them).

The compiler will then garbage-collect the init/shutdown functions (and
all related code) of the disabled Maple drivers.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* Add weak init/shutdown of the whole Maple bus

If none of the Maple drivers have been enabled, we can skip
initialization/shutdown of the whole Maple bus, allowing the compiler to
garbage-collect more code.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* Enable VMU FS according to init flags

Use weak init/shutdown functions for the VMU FS. This allows the VMU FS
code to be automatically garbage-collected if the INIT_VMU init flag has
not been set (it is set by default).

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

---------

Signed-off-by: Paul Cercueil <paul@crapouillou.net>