Threaded ASIC IRQ handlers (#393)
* irq: Use standard stdint.h types

Use the standard 'uint32_t' type instead of the custom 'uint32' one.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* dc: irq: Pass data pointer to IRQ handlers

Update the IRQ API to allow clients to register a data pointer alongside
the callback.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* dc: asic: Add function asic_evt_remove_handler()

This function can be used to unregister any handler set to the given
ASIC event.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* dc: asic: Pass data pointer to IRQ handlers

Allow registering a pointer alongside the IRQ handler. The pointer will
be passed as argument to the callback function.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* dc: vblank: Pass data pointer to VBLANK IRQ handler

Allow registering a pointer alongside the IRQ handler. The pointer will
be passed as argument to the callback function.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* maple: Pass down state pointer to IRQ callbacks

Avoid global accesses to the 'maple_state' variable; instead, access it
through a 'state' parameter that's passed as argument to the IRQ
callbacks.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* dc: ubc: Pass data pointer to IRQ handlers

Update the UBC interrupt handler code to pass a pointer to the
channel_state array.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* dc: asic: Add support for threaded IRQ handlers

Threaded IRQ handlers are functions which will be called within a
thread, when a given IRQ happens.

The advantage vs. regular IRQ handlers is that they are executed outside
the IRQ context, and as a result, can hold locks and sleep if needed.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

---------

Signed-off-by: Paul Cercueil <paul@crapouillou.net>