Worker threads (#505)
* thread: Add worker threads

Worker threads are threads that are associated to a given work function.
When notified (through thd_worker_notify()), the worker thread will call
its work function.

This is useful for implementing asynchronous function calls. It is
especially useful to implement threaded interrupt handlers.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* asic: Use worker thread to implement threaded interrupt handlers

Most of the threaded interrupt handler code was duplicated vs. the
worker thread code. Update the ASIC code to now use the worker thread
API instead.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* thread: Add optional jobs API for use with worker threads

Using this API, the program can enqueue specific jobs to the worker
threads, and dequeue them in their respective work function, in a FIFO
fashion.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* thread: Add explicit constructor for worker threads

Add thd_worker_create_ex() which takes an extra parameter that can be
used to specify the thread's attributes.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>

* Update include/kos/worker_thread.h

---------

Signed-off-by: Paul Cercueil <paul@crapouillou.net>
Co-authored-by: Falco Girgis <gyrovorbis@gmail.com>
Co-authored-by: darcagn <darcagn@protonmail.com>